+++ src/hatch/publish/index.py
@@ -5,21 +5,7 @@ from typing import Generator
 
 from hatch.publish.plugin.interface import PublisherInterface
 from hatch.utils.fs import Path
-
-MULTIPLE_USE_METADATA_FIELDS = {
-    'classifier',
-    'dynamic',
-    'license_file',
-    'obsoletes_dist',
-    'platform',
-    'project_url',
-    'provides_dist',
-    'provides_extra',
-    'requires_dist',
-    'requires_external',
-    'supported_platform',
-}
-RENAMED_METADATA_FIELDS = {'classifier': 'classifiers', 'project_url': 'project_urls'}
+from hatchling.metadata.utils import normalize_project_name
 
 
 class IndexPublisher(PublisherInterface):
@@ -36,12 +22,12 @@ class IndexPublisher(PublisherInterface):
         """
         https://warehouse.readthedocs.io/api-reference/legacy.html#upload-api
         """
-        import hashlib
-        import io
         from collections import defaultdict
-        from urllib.parse import urlparse
 
         import httpx
+
+        from hatch.index.core import PackageIndex
+        from hatch.index.publish import get_sdist_form_data, get_wheel_form_data
 
         if not artifacts:
             from hatchling.builders.constants import DEFAULT_BUILD_DIRECTORY
@@ -56,6 +42,8 @@ class IndexPublisher(PublisherInterface):
         if repo in self.repos:
             repo = self.repos[repo]
 
+        index = PackageIndex(repo)
+
         cached_user_file = CachedUserFile(self.cache_dir)
         updated_user = None
         if 'user' in options:
@@ -69,6 +57,7 @@ class IndexPublisher(PublisherInterface):
                         self.app.abort('Missing required option: user')
                     else:
                         user = updated_user = self.app.prompt('Enter your username')
+        index.user = user
 
         updated_auth = None
         if 'auth' in options:
@@ -84,13 +73,8 @@ class IndexPublisher(PublisherInterface):
                         self.app.abort('Missing required option: auth')
                     else:
                         auth = updated_auth = self.app.prompt('Enter your credentials', hide_input=True)
+        index.auth = auth
 
-        repo_components = urlparse(repo)
-        domain = repo_components.netloc
-        if domain == 'upload.pypi.org':  # no cov
-            domain = 'pypi.org'
-
-        index_url = f'{repo_components.scheme}://{domain}/simple/'
         existing_artifacts: dict[str, set[str]] = {}
 
         # Use as an ordered set
@@ -99,9 +83,9 @@ class IndexPublisher(PublisherInterface):
         artifacts_found = False
         for artifact in recurse_artifacts(artifacts, self.root):
             if artifact.name.endswith('.whl'):
-                data = get_wheel_form_data(self.app, artifact)
+                data = get_wheel_form_data(artifact)
             elif artifact.name.endswith('.tar.gz'):
-                data = get_sdist_form_data(self.app, artifact)
+                data = get_sdist_form_data(artifact)
             else:
                 continue
 
@@ -121,7 +105,7 @@ class IndexPublisher(PublisherInterface):
             project_name = normalize_project_name(data['name'])
             if project_name not in existing_artifacts:
                 try:
-                    response = httpx.get(f'{index_url}{project_name}/')
+                    response = httpx.get(str(index.urls.simple.child(project_name, '')))
                     response.raise_for_status()
                 except Exception:  # no cov
                     existing_artifacts[project_name] = set()
@@ -132,63 +116,27 @@ class IndexPublisher(PublisherInterface):
                 self.app.display_warning('already exists')
                 continue
 
-            data[':action'] = 'file_upload'
-            data['protocol_version'] = '1'
+            try:
+                index.upload_artifact(artifact, data)
+            except Exception as e:
+                self.app.display_error('failed')
+                self.app.abort(str(e).replace(index.auth, '*****'))
+            else:
+                self.app.display_success('success')
 
-            with artifact.open('rb') as f:
-                # https://github.com/pypa/warehouse/blob/7fc3ce5bd7ecc93ef54c1652787fb5e7757fe6f2/tests/unit/packaging/test_tasks.py#L189-L191
-                md5_hash = hashlib.md5()  # nosec B303
-                sha256_hash = hashlib.sha256()
-                blake2_256_hash = hashlib.blake2b(digest_size=32)
-
-                while True:
-                    chunk = f.read(io.DEFAULT_BUFFER_SIZE)
-                    if not chunk:
-                        break
-
-                    md5_hash.update(chunk)
-                    sha256_hash.update(chunk)
-                    blake2_256_hash.update(chunk)
-
-                data['md5_digest'] = md5_hash.hexdigest()
-                data['sha256_digest'] = sha256_hash.hexdigest()
-                data['blake2_256_digest'] = blake2_256_hash.hexdigest()
-
-                f.seek(0)
-
-                try:
-                    response = httpx.post(
-                        repo,
-                        data=data,
-                        files={'content': (artifact.name, f, 'application/octet-stream')},
-                        auth=(user, auth),
-                    )
-                    response.raise_for_status()
-                except Exception as e:
-                    self.app.display_error('failed')
-                    self.app.abort(str(e).replace(auth, '*****'))
-                else:
-                    self.app.display_success('success')
-
-                    existing_artifacts[project_name].add(artifact.name)
-                    project_versions[project_name][data['version']] = None
+                existing_artifacts[project_name].add(artifact.name)
+                project_versions[project_name][data['version']] = None
 
         if not artifacts_found:
             self.app.abort('No artifacts found')
         elif not project_versions:
             self.app.abort(code=0)
 
-        if domain.endswith('pypi.org'):
-            for project_name, versions in project_versions.items():
-                self.app.display_info()
-                self.app.display_mini_header(project_name)
-                for version in versions:
-                    self.app.display_info(f'https://{domain}/project/{project_name}/{version}/')
-        else:  # no cov
-            for project_name in project_versions:
-                self.app.display_info()
-                self.app.display_mini_header(project_name)
-                self.app.display_info(f'{index_url}{project_name}/')
+        for project_name, versions in project_versions.items():
+            self.app.display_info()
+            self.app.display_mini_header(project_name)
+            for version in versions:
+                self.app.display_info(str(index.urls.project.child(project_name, version, '').to_iri()))
 
         if updated_user is not None:
             cached_user_file.set_user(repo, user)
@@ -197,92 +145,6 @@ class IndexPublisher(PublisherInterface):
             import keyring
 
             keyring.set_password(repo, user, auth)
-
-
-def get_wheel_form_data(app, artifact):
-    import zipfile
-
-    from packaging.tags import parse_tag
-
-    with zipfile.ZipFile(str(artifact), 'r') as zip_archive:
-        dist_info_dir = ''
-        for path in zip_archive.namelist():
-            root = path.split('/', 1)[0]
-            if root.endswith('.dist-info'):
-                dist_info_dir = root
-                break
-        else:  # no cov
-            app.abort(f'Could not find the `.dist-info` directory in wheel: {artifact}')
-
-        try:
-            with zip_archive.open(f'{dist_info_dir}/METADATA') as zip_file:
-                metadata_file_contents = zip_file.read().decode('utf-8')
-        except KeyError:  # no cov
-            app.abort(f'Could not find a `METADATA` file in the `{dist_info_dir}` directory')
-        else:
-            data = parse_headers(metadata_file_contents)
-
-    data['filetype'] = 'bdist_wheel'
-
-    # Examples:
-    # cryptography-3.4.7-pp37-pypy37_pp73-manylinux2014_x86_64.whl -> pp37
-    # hatchling-1rc1-py2.py3-none-any.whl -> py2.py3
-    tag_component = '-'.join(artifact.stem.split('-')[-3:])
-    data['pyversion'] = '.'.join(sorted({tag.interpreter for tag in parse_tag(tag_component)}))
-
-    return data
-
-
-def get_sdist_form_data(app, artifact):
-    import tarfile
-
-    with tarfile.open(str(artifact), 'r:gz') as tar_archive:
-        pkg_info_dir_parts = []
-        for tar_info in tar_archive:
-            if tar_info.isfile():
-                pkg_info_dir_parts.append(tar_info.name.split('/', 1)[0])
-                break
-            else:  # no cov
-                pass
-        else:  # no cov
-            app.abort(f'Could not find any files in sdist: {artifact}')
-
-        pkg_info_dir_parts.append('PKG-INFO')
-        pkg_info_path = '/'.join(pkg_info_dir_parts)
-        try:
-            with tar_archive.extractfile(pkg_info_path) as tar_file:
-                metadata_file_contents = tar_file.read().decode('utf-8')
-        except KeyError:  # no cov
-            app.abort(f'Could not find file: {pkg_info_path}')
-        else:
-            data = parse_headers(metadata_file_contents)
-
-    data['filetype'] = 'sdist'
-    data['pyversion'] = 'source'
-
-    return data
-
-
-def parse_headers(metadata_file_contents):
-    import email
-
-    message = email.message_from_string(metadata_file_contents)
-
-    headers = {'description': message.get_payload()}
-
-    for header, value in message.items():
-        normalized_header = header.lower().replace('-', '_')
-        header_name = RENAMED_METADATA_FIELDS.get(normalized_header, normalized_header)
-
-        if normalized_header in MULTIPLE_USE_METADATA_FIELDS:
-            if header_name in headers:
-                headers[header_name].append(value)
-            else:
-                headers[header_name] = [value]
-        else:
-            headers[header_name] = value
-
-    return headers
 
 
 def recurse_artifacts(artifacts: list, root) -> Generator[Path, None, None]:
@@ -295,11 +157,6 @@ def recurse_artifacts(artifacts: list, root) -> Generator[Path, None, None]:
             yield artifact
         elif artifact.is_dir():
             yield from artifact.iterdir()
-
-
-def normalize_project_name(name):
-    # https://peps.python.org/pep-0503/#normalized-names
-    return re.sub(r'[-_.]+', '-', name).lower()
 
 
 def parse_artifacts(artifact_payload):