+++ docs/history.md
@@ -14,8 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Values for environment `env-vars` now support context formatting
 - Add `name` override for environments to allow for regular expression matching
+- The `index` publisher now better supports non-PyPI indices
 - Display help text the first time the `shell` command is executed
 - Add support for Almquist (`ash`) shells
+- Add `hyperlink` as a dependency for better handling of package index URLs
 - Bump the minimum supported version of `virtualenv` to 20.16.2
 
 ***Fixed:***